Create dos.py
This project is a simple simulation of a Denial of Service (DoS) attack using Python. The goal of the project is to provide a simple simulation of a DoS attack at a low level using the UDP protocol. It continuously sends packets to a specified target over the internet or local network in an attempt to interrupt the target's server response